Title: 화재 진압, Time: 312ms, Memory: 18MB, Status: Wrong Answer - Codetree

+++ 250517/화재 진압/fire-suppression.java
@@ -24,7 +24,7 @@ public class Main {
             stationPos[i] = Integer.parseInt(st.nextToken());
             s.add(stationPos[i]);
         }
-        int ans = 0;
+        long ans = 0;
 
         for(int i = 0; i < n; i++) {
             Integer x1 = s.ceiling(firePos[i]);
@@ -33,9 +33,9 @@ public class Main {
             // if(x1 != null && x2 != null)
             //     ans = Math.max(Math.abs(firePos[i]- x1) , Math.abs(firePos[i]- x2));
             if(x1 != null)
-                ans = Math.max(ans, Math.abs(firePos[i] - x1));
+                ans = Math.max(ans, Math.abs(firePos[i] - (long)x1));
             if(x2 != null)
-                ans = Math.max(ans , Math.abs(firePos[i] - x2));
+                ans = Math.max(ans , Math.abs(firePos[i] - (long)x2));
         }
 
         System.out.print(ans);
